# Fix Int overrun potential in dict skip_deleted