Previously there was the potential for an infinite loop in the `skip_deleted` function in `L=typemax(Int)`.   This fixes that and provides some performance benefits.

Baseline:
```julia
julia> using BenchmarkTools

julia> d=Dict(v=>v for v in 1:1000);

julia> @benchmark collect(values(d))
BenchmarkTools.Trial: 
  memory estimate:  7.95 KiB
  allocs estimate:  2
  --------------
  minimum time:     6.313 μs (0.00% GC)
  median time:      8.026 μs (0.00% GC)
  mean time:        11.180 μs (15.87% GC)
  maximum time:     12.011 ms (99.82% GC)
  --------------
  samples:          10000
  evals/sample:     5

julia> @benchmark sum(values(d))
BenchmarkTools.Trial: 
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     5.716 μs (0.00% GC)
  median time:      6.483 μs (0.00% GC)
  mean time:        8.101 μs (0.00% GC)
  maximum time:     1.149 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     6

julia> @benchmark sum(v->v.second,d)
BenchmarkTools.Trial: 
  memory estimate:  272 bytes
  allocs estimate:  7
  --------------
  minimum time:     8.236 μs (0.00% GC)
  median time:      10.181 μs (0.00% GC)
  mean time:        12.392 μs (0.00% GC)
  maximum time:     1.016 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     3
```

After fix:
```julia
julia> using BenchmarkTools

julia> d=Dict(v=>v for v in 1:1000);

julia> @benchmark collect(values(d))
BenchmarkTools.Trial: 
  memory estimate:  7.95 KiB
  allocs estimate:  2
  --------------
  minimum time:     4.154 μs (0.00% GC)
  median time:      7.265 μs (0.00% GC)
  mean time:        14.121 μs (17.00% GC)
  maximum time:     12.819 ms (99.87% GC)
  --------------
  samples:          10000
  evals/sample:     7

julia> @benchmark sum(values(d))
BenchmarkTools.Trial: 
  memory estimate:  272 bytes
  allocs estimate:  8
  --------------
  minimum time:     5.441 μs (0.00% GC)
  median time:      5.808 μs (0.00% GC)
  mean time:        6.910 μs (0.00% GC)
  maximum time:     243.749 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     6

julia> @benchmark sum(v->v.second,d)
BenchmarkTools.Trial: 
  memory estimate:  272 bytes
  allocs estimate:  7
  --------------
  minimum time:     5.749 μs (0.00% GC)
  median time:      6.785 μs (0.00% GC)
  mean time:        10.137 μs (0.00% GC)
  maximum time:     2.498 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     6
```