@JeffBezanson 2019-03-25
Any speedup is nice of course, but do I understand right that this is guarding against arrays of length `typemax(Int)`? It doesn't seem necessary to worry about that.