@vchuravy 2019-03-25
> against arrays of length typemax(Int)? It doesn't seem necessary to worry about that.

Problem is that LLVM goes and proves an overflow and then decides that this loop is not worthy of future optimization.