@yuyichao 2019-03-25
LLVM never seems to clone loop for this purpose. Even though it almost always emit the scalar loop anyway. Saw this for both overflow problem and alignment....