@ndinsmore 2019-03-25
@JeffBezanson some relevant discussion happened in #31416. 
 
For me this is for the burndown of a list of things that could get into the way of higher level iterators getting in the way of being really lean loops.     In digging into that and trying to understanding what stopped fast loops from being emitted this showed up and I thought it should just be fixed while still top of mind.