@ndinsmore 2019-03-25
A more poweful benchmark might be this
```julia
using BenchmarkTools
d=Dict(v=>v for v in 1:1000)


function for_sum(f,iter)    
    ret=0
    for v in iter
        ret += f(v)
    end
    return ret
end

function while_sum(f,iter)    
    ret=0
    next = iterate(iter)
    while next !== nothing
        (v, state) = next
        ret+=f(v)
        next = iterate(iter, state)
    end
    return ret
end
```
Before:
```julia
julia> @benchmark for_sum(v->v,values(d))
BenchmarkTools.Trial: 
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     6.816 μs (0.00% GC)
  median time:      7.893 μs (0.00% GC)
  mean time:        9.323 μs (0.00% GC)
  maximum time:     294.384 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     4

julia> @benchmark while_sum(v->v,values(d))
BenchmarkTools.Trial: 
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     5.534 μs (0.00% GC)
  median time:      6.920 μs (0.00% GC)
  mean time:        8.478 μs (0.00% GC)
  maximum time:     544.237 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     5

```
after:
```Julia
julia> @benchmark for_sum(v->v,values(d))
BenchmarkTools.Trial: 
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     3.461 μs (0.00% GC)
  median time:      4.886 μs (0.00% GC)
  mean time:        5.145 μs (0.00% GC)
  maximum time:     176.057 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     8

julia> @benchmark while_sum(v->v,values(d))
BenchmarkTools.Trial: 
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     3.501 μs (0.00% GC)
  median time:      3.723 μs (0.00% GC)
  mean time:        4.421 μs (0.00% GC)
  maximum time:     127.837 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     8
```
These benchmark tailored to be like use in the wild.