@ndinsmore 2019-03-29
This provides a good performance bump is there anything that is blocking?